+++ packages/deploy/src/bff/launch-routes.ts
@@ -16,25 +16,165 @@
 //
 // No GitHub credential is configured here: CI status and rerun go through the
 // Aomi backend, whose GitHub App installation token makes every GitHub call.
+//
+// Also owns launch config resolution (env → LaunchConfig) and deployment
+// payload → client summary mappers.
 // =============================================================================
 
 import type { BackendClient } from "../backend";
-import type { UserProject } from "../types";
+import type { DeployPayload, UserProject } from "../types";
 import { assertServerOnly } from "../backend";
-import { resolveLaunchConfig, type LaunchConfig } from "./config";
+import { DEFAULT_TEMPLATE_REPO } from "../launch/contracts";
 import { launchErrorResponse } from "./errors";
-import { appNamesFromDeployment, releaseTagsFromDeployment } from "./mappers";
-import { createDefaultGuards, jsonResponse, type LaunchGuards } from "./guards";
-import { missingSecretsForActivation } from "./release-manifest";
-import { randomHex } from "./cookies";
-import type { GitHubSession } from "./github-session";
 import {
-  isValidProjectId,
+  createDefaultGuards,
   isValidDeploymentId,
   isValidInstallationId,
+  isValidProjectId,
   isValidReleaseTags,
   isValidRepo,
-} from "./validate";
+  jsonResponse,
+  randomHex,
+  type LaunchGuards,
+} from "./http";
+import type { GitHubSession } from "./auth";
+import { missingSecretsForActivation } from "./release-manifest";
+
+export const DEFAULT_DEPLOY_PLATFORM = "community";
+export { DEFAULT_TEMPLATE_REPO };
+
+export type LaunchConfig = {
+  /** Platform every launch route operates on (first of `platforms`). */
+  platform: string;
+  /** All deployable platforms, first is the default. */
+  platforms: string[];
+  /** Platforms whose apps show in the catalog (optional). */
+  catalogPlatforms: string[];
+  /** Template `owner/repo` the one-shot flow forks. */
+  templateRepo: string;
+  /** Create scaffolded repos as private. */
+  createdRepoPrivate: boolean;
+  /** Target tags applied on activation (e.g. ["staging"]). */
+  targetTags: string[];
+};
+
+function envString(name: string, fallback: string): string {
+  return process.env[name]?.trim() || fallback;
+}
+
+function envBoolean(name: string, fallback: boolean): boolean {
+  const value = process.env[name]?.trim().toLowerCase();
+  if (!value) return fallback;
+  return value === "1" || value === "true" || value === "yes";
+}
+
+function envList(name: string): string[] {
+  return (process.env[name] ?? "")
+    .split(",")
+    .map((value) => value.trim())
+    .filter(Boolean);
+}
+
+function dedupe(values: string[]): string[] {
+  return Array.from(new Set(values));
+}
+
+function envJsonOrCommaList(name: string): string[] {
+  const raw = process.env[name]?.trim();
+  if (!raw) return [];
+
+  try {
+    const parsed = JSON.parse(raw) as unknown;
+    if (Array.isArray(parsed)) {
+      return dedupe(
+        parsed
+          .map((value) => (typeof value === "string" ? value.trim() : ""))
+          .filter(Boolean),
+      );
+    }
+  } catch {
+    // Fall through to comma-separated parsing for Vercel/plain .env ergonomics.
+  }
+
+  return dedupe(
+    raw
+      .split(",")
+      .map((value) => value.trim())
+      .filter(Boolean),
+  );
+}
+
+function deployPlatformsFromEnv(): string[] {
+  for (const name of [
+    "APP_DEPLOY_PLATFORMS",
+    "NEXT_PUBLIC_APP_DEPLOY_PLATFORMS",
+    "APP_DEPLOY_PLATFORM",
+    "NEXT_PUBLIC_APP_DEPLOY_PLATFORM",
+  ]) {
+    const platforms = envJsonOrCommaList(name);
+    if (platforms.length > 0) return platforms;
+  }
+
+  return [DEFAULT_DEPLOY_PLATFORM];
+}
+
+function catalogPlatformsFromEnv(): string[] {
+  for (const name of [
+    "APP_CATALOG_PLATFORMS",
+    "NEXT_PUBLIC_APP_CATALOG_PLATFORMS",
+  ]) {
+    const platforms = envJsonOrCommaList(name);
+    if (platforms.length > 0) return platforms;
+  }
+
+  return [];
+}
+
+/** Resolve the launch config: explicit overrides win, then env, then defaults. */
+export function resolveLaunchConfig(
+  overrides?: Partial<LaunchConfig>,
+): LaunchConfig {
+  const platforms =
+    overrides?.platforms ??
+    (overrides?.platform ? [overrides.platform] : deployPlatformsFromEnv());
+
+  return {
+    platform: overrides?.platform ?? platforms[0],
+    platforms,
+    catalogPlatforms: overrides?.catalogPlatforms ?? catalogPlatformsFromEnv(),
+    templateRepo:
+      overrides?.templateRepo ??
+      (process.env.APP_DEPLOY_TEMPLATE_REPO?.trim() ||
+        envString("NEXT_PUBLIC_APP_DEPLOY_TEMPLATE_REPO", DEFAULT_TEMPLATE_REPO)),
+    createdRepoPrivate:
+      overrides?.createdRepoPrivate ??
+      envBoolean("APP_DEPLOY_CREATED_REPO_PRIVATE", false),
+    targetTags: overrides?.targetTags ?? envList("APP_DEPLOY_TARGET_TAGS"),
+  };
+}
+
+// Tolerant of both snake_case (raw backend) and camelCase (BackendClient)
+// app records, since status payloads have carried both shapes.
+type AppsLike = { platform?: { apps?: Array<Record<string, unknown>> } };
+
+export function releaseTagsFromDeployment(
+  deployment?: DeployPayload | AppsLike,
+): string[] {
+  const apps = (deployment as AppsLike | undefined)?.platform?.apps ?? [];
+  return apps
+    .map((app) => (app.release_tag ?? app.releaseTag) as string | undefined)
+    .map((tag) => tag?.trim())
+    .filter((tag): tag is string => Boolean(tag));
+}
+
+export function appNamesFromDeployment(
+  deployment?: DeployPayload | AppsLike,
+): string[] {
+  const apps = (deployment as AppsLike | undefined)?.platform?.apps ?? [];
+  return apps
+    .map((app) => (app.name as string | undefined)?.trim())
+    .filter((name): name is string => Boolean(name));
+}
 
 export type LaunchRouteHandler = (req: Request) => Promise<Response>;
