+++ packages/deploy/test/launch-connection-result.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { connectionResult } from "../src/launch/connection-result";
+import { connectionResult } from "../src/launch/state";
 
 describe("connectionResult", () => {
   it("reports a bound repository as connected", () => {
